+++ src/Language/LanguageIn.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Fisharebest\Localization\Language;
+
+/**
+ * Class LanguageIn - Representation of the Indonesian language (deprecated).
+ *
+ * @author    Greg Roach <greg@subaqua.co.uk>
+ * @copyright (c) 2022 Greg Roach
+ * @license   GPL-3.0-or-later
+ */
+class LanguageIn extends LanguageId
+{
+}