+++ src/Language/LanguageJi.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Fisharebest\Localization\Language;
+
+/**
+ * Class LanguageJi - Representation of the Yiddish language (deprecated).
+ *
+ * @author    Greg Roach <greg@subaqua.co.uk>
+ * @copyright (c) 2022 Greg Roach
+ * @license   GPL-3.0-or-later
+ */
+class LanguageJi extends LanguageYi
+{
+}