+++ src/Language/LanguageJw.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Fisharebest\Localization\Language;
+
+/**
+ * Class LanguageJw - Representation of the Javanese language (deprecated).
+ *
+ * @author    Greg Roach <greg@subaqua.co.uk>
+ * @copyright (c) 2022 Greg Roach
+ * @license   GPL-3.0-or-later
+ */
+class LanguageJw extends LanguageJv
+{
+}