+++ src/Language/LanguageMo.php
@@ -6,26 +6,12 @@ use Fisharebest\Localization\PluralRule\PluralRule5;
 use Fisharebest\Localization\Territory\TerritoryMd;
 
 /**
- * Class LanguageMo - Representation of the Moldovan language.
+ * Class LanguageMo - Representation of the Moldovan language (deprecated).
  *
  * @author    Greg Roach <greg@subaqua.co.uk>
  * @copyright (c) 2022 Greg Roach
  * @license   GPL-3.0-or-later
  */
-class LanguageMo extends AbstractLanguage implements LanguageInterface
+class LanguageMo extends LanguageRo
 {
-    public function code()
-    {
-        return 'mo';
-    }
-
-    public function defaultTerritory()
-    {
-        return new TerritoryMd();
-    }
-
-    public function pluralRule()
-    {
-        return new PluralRule5();
-    }
 }