+++ src/Locale/LocaleJi.php
@@ -3,12 +3,12 @@
 namespace Fisharebest\Localization\Locale;
 
 /**
- * Class LocaleMo - Moldovan (deprecated).
+ * Class LocaleJi - Yiddish (deprecated)
  *
  * @author    Greg Roach <greg@subaqua.co.uk>
  * @copyright (c) 2022 Greg Roach
  * @license   GPL-3.0-or-later
  */
-class LocaleMo extends LocaleRo
+class LocaleJi extends LocaleYi
 {
 }