+++ test/Language/LanguageIwTest.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Fisharebest\LocalizationTest\Language;
+
+use Fisharebest\Localization\Language\LanguageIw;
+use Fisharebest\Localization\PluralRule\PluralRule1;
+use Fisharebest\Localization\Script\ScriptHebr;
+use Fisharebest\Localization\Territory\TerritoryIl;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @author    Greg Roach <greg@subaqua.co.uk>
+ * @copyright (c) 2022 Greg Roach
+ * @license   GPL-3.0-or-later
+ *
+ * @covers \Fisharebest\Localization\Language\AbstractLanguage
+ * @covers \Fisharebest\Localization\Language\LanguageHe
+ */
+class LanguageIwTest extends TestCase
+{
+    /**
+     * Test the properties
+     */
+    public function testProperties()
+    {
+        $language = new LanguageIw(); // iw is deprecated - use he
+
+        self::assertEquals(new ScriptHebr(), $language->defaultScript());
+        self::assertEquals(new TerritoryIl(), $language->defaultTerritory());
+        self::assertEquals(new PluralRule1(), $language->pluralRule());
+        self::assertSame('he', $language->code());
+    }
+}