+++ test/Language/LanguageJiTest.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Fisharebest\LocalizationTest\Language;
+
+use Fisharebest\Localization\Language\LanguageJi;
+use Fisharebest\Localization\PluralRule\PluralRule1;
+use Fisharebest\Localization\Script\ScriptHebr;
+use Fisharebest\Localization\Territory\Territory001;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @author    Greg Roach <greg@subaqua.co.uk>
+ * @copyright (c) 2022 Greg Roach
+ * @license   GPL-3.0-or-later
+ *
+ * @covers \Fisharebest\Localization\Language\AbstractLanguage
+ * @covers \Fisharebest\Localization\Language\LanguageYi
+ */
+class LanguageJiTest extends TestCase
+{
+    /**
+     * Test the properties
+     */
+    public function testProperties()
+    {
+        $language = new LanguageJi(); // ji is deprecated - use yi
+
+        self::assertEquals(new ScriptHebr(), $language->defaultScript());
+        self::assertEquals(new Territory001(), $language->defaultTerritory());
+        self::assertEquals(new PluralRule1(), $language->pluralRule());
+        self::assertSame('yi', $language->code());
+    }
+}