+++ test/Language/LanguageJwTest.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Fisharebest\LocalizationTest\Language;
+
+use Fisharebest\Localization\Language\LanguageJw;
+use Fisharebest\Localization\PluralRule\PluralRule0;
+use Fisharebest\Localization\Script\ScriptJava;
+use Fisharebest\Localization\Territory\TerritoryId;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @author    Greg Roach <greg@subaqua.co.uk>
+ * @copyright (c) 2022 Greg Roach
+ * @license   GPL-3.0-or-later
+ *
+ * @covers \Fisharebest\Localization\Language\AbstractLanguage
+ * @covers \Fisharebest\Localization\Language\LanguageJv
+ */
+class LanguageJwTest extends TestCase
+{
+    /**
+     * Test the properties
+     */
+    public function testProperties()
+    {
+        $language = new LanguageJw(); // jw is deprecated - use jv
+
+        self::assertEquals(new ScriptJava(), $language->defaultScript());
+        self::assertEquals(new TerritoryId(), $language->defaultTerritory());
+        self::assertEquals(new PluralRule0(), $language->pluralRule());
+        self::assertSame('jv', $language->code());
+    }
+}