+++ test/Locale/LocaleMoTest.php
@@ -2,11 +2,11 @@
 
 namespace Fisharebest\LocalizationTest\Locale;
 
-use Fisharebest\Localization\Language\LanguageMo;
+use Fisharebest\Localization\Language\LanguageRo;
 use Fisharebest\Localization\Locale\LocaleMo;
 use Fisharebest\Localization\PluralRule\PluralRule5;
 use Fisharebest\Localization\Script\ScriptLatn;
-use Fisharebest\Localization\Territory\TerritoryMd;
+use Fisharebest\Localization\Territory\TerritoryRo;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -17,7 +17,7 @@ use PHPUnit\Framework\TestCase;
  * @license   GPL-3.0-or-later
  *
  * @covers \Fisharebest\Localization\Locale\AbstractLocale
- * @covers \Fisharebest\Localization\Locale\LocaleMo
+ * @covers \Fisharebest\Localization\Locale\LocaleRo
  */
 class LocaleMoTest extends TestCase
 {
@@ -26,21 +26,21 @@ class LocaleMoTest extends TestCase
      */
     public function testProperties()
     {
-        $locale = new LocaleMo();
+        $locale = new LocaleMo(); // mo is deprecated - use ro
 
-        self::assertEquals(new LanguageMo(), $locale->language());
+        self::assertEquals(new LanguageRo(), $locale->language());
         self::assertEquals(new PluralRule5(), $locale->pluralRule());
         self::assertEquals(new ScriptLatn(), $locale->script());
-        self::assertEquals(new TerritoryMd(), $locale->territory());
+        self::assertEquals(new TerritoryRo(), $locale->territory());
         self::assertNull($locale->variant());
-        self::assertSame('mo_MD', $locale->code());
-        self::assertSame('unicode_ci', $locale->collation());
+        self::assertSame('ro_RO', $locale->code());
+        self::assertSame('romanian_ci', $locale->collation());
         self::assertSame('0123456789', $locale->digits('0123456789'));
         self::assertSame('ltr', $locale->direction());
-        self::assertSame('limba moldovenească', $locale->endonym());
-        self::assertSame('LIMBA MOLDOVENEASCĂ', $locale->endonymSortable());
-        self::assertSame('lang="mo"', $locale->htmlAttributes());
-        self::assertSame('mo', $locale->languageTag());
+        self::assertSame('română', $locale->endonym());
+        self::assertSame('ROMANA', $locale->endonymSortable());
+        self::assertSame('lang="ro"', $locale->htmlAttributes());
+        self::assertSame('ro', $locale->languageTag());
         self::assertSame('-123', $locale->number(-123));
         self::assertSame('12.345.678,09', $locale->number(12345678.09));
         self::assertSame('1.234,56 %', $locale->percent(12.3456));